requirements.txt
PyQt5>=5.15.0
Pillow>=9.0.0
appdirs>=1.4.0
PyInstaller>=5.0.0